Split tests to separate files
