+++ django_site/core/tests/__init__.py
@@ -1,0 +1,5 @@
+"""Test package for the core app.
+
+Split by feature area (test_auth, test_orcid, test_papers, ...).
+Run the whole suite with: python manage.py test core
+"""
